# Memoize parseUserAgent