Hi Statsig! Submitting an optimization here that would be a fairly large improvement for us. Let me know if you have any questions!

This PR memoizes the most recent call to `parseUserAgent`. This function is potentially called many times (depending on the number and type of rules used) per `getClientInitializeResponse` request with the same user agent, and can be quite expensive to run.

Locally for our use case, I see around a 40% decrease in `getClientInitializeResponse` time:

| Before | After |
|---|---|
| 20ms | 12ms | 
| <img width="1789" alt="Screenshot 2024-10-11 at 17 21 57" src="https://github.com/user-attachments/assets/dac394b5-a5f9-4510-83fe-20d0098eef3f"> | <img width="1789" alt="Screenshot 2024-10-11 at 17 22 28" src="https://github.com/user-attachments/assets/fced3b7c-a299-4a39-a203-419f32e49a62"> |