@tore-statsig 2024-10-19
released in https://github.com/statsig-io/node-js-server-sdk/releases/tag/5.29.0